Release 950109
Thu Jan  5 13:37:42 1995  Cameron Heide  (heide@ee.ualberta.ca)

	* [memory/global.c]
	GlobalCompact should now return the correct value when the
	largest run of free blocks includes the last block.

	* [windows/mdi.c]
	Tiling and cascading windows without any MDI children should
	no longer crash (assuming no-op is the correct thing to do).

Sun Jan  1 23:30:25 1995  Fons Botman  <botman@rabo.nl>

	* [objects/font.c]
	GetTextExtentPoint: fixed debug output, str is counted string, not
        zero terminated.

	* [if1632/relay.c]
	DLLRelay: when debugging_stack got segv, added upper bound for
        stack dump.

+++ miscemu/int25.c
@@ -25,7 +25,7 @@ int do_int25(struct sigcontext_struct *context)
 		length = CX;
 	}
 	dprintf_int(stdnimp, "int25: abs diskread, drive %d, sector %ld, "
-	"count %ld, buffer %d\n", EAX & 0xff, begin, length, (int) dataptr);
+	"count %ld, buffer %d\n", (int)EAX & 0xff, begin, length, (int) dataptr);
 
 	memset(dataptr, 0, length * 512);
